Fix OPAM metadata
This PR removes lwt.unix (since that's not an OPAM package name but a
findlib package name that is part of the lwt OPAM package) and also
regenerates the opam file thus updates it to the current state of
dune-project.

This change is necessary as Dune 3.13 introduces strict checking for
OPAM package names and this package fails when applying these checks.